move data prepration stuff to experimental_helper.data
