+++ cezo_fl/fl_helpers.py
@@ -1,2 +1,6 @@
 def get_client_name(client_index: int) -> str:
     return f"client{client_index}"
+
+
+def get_server_name() -> str:
+    return "server"
