+++ decomfl_main.py
@@ -9,7 +9,7 @@ from tqdm import tqdm
 from byzantine import aggregation as byz_agg
 from byzantine import attack as byz_attack
 from cezo_fl.client import ResetClient
-from cezo_fl.fl_helpers import get_client_name
+from cezo_fl.fl_helpers import get_client_name, get_server_name
 from cezo_fl.server import CeZO_Server
 from cezo_fl.util import model_helpers
 from experiment_helper import prepare_settings
@@ -22,8 +22,8 @@ from experiment_helper.cli_parser import (
     FederatedLearningSetting,
     RGESetting,
 )
-
-from preprocess import preprocess
+from experiment_helper.device import use_device
+from experiment_helper.data import get_dataloaders
 
 
 class CliSetting(
@@ -71,7 +71,7 @@ def setup_server_and_clients(
         )
         clients.append(client)
 
-    server_device = device_map["server"]
+    server_device = device_map[get_server_name()]
     server = CeZO_Server(
         clients,
         server_device,
@@ -145,8 +145,8 @@ def setup_server_and_clients(
 if __name__ == "__main__":
     args = CliSetting()
     print(args)
-    device_map, train_loaders, test_loader = preprocess(args)
-
+    device_map = use_device(args, args.num_clients)
+    train_loaders, test_loader = get_dataloaders(args, args.num_clients)
     server = setup_server_and_clients(args, device_map, train_loaders)
 
     if args.log_to_tensorboard:
